+++ src/variables/regions.tsx
